- add setup.py architecture (exports "spider" entrypoint) - cleanup internal architecture/imports - aggressive python 3.6+ conversion - no promises I didn't break everything...
