+++ htcondor_es/queues.py
@@ -11,11 +11,7 @@ import multiprocessing
 
 import htcondor
 
-import htcondor_es.es
-from htcondor_es.utils import send_email_alert, time_remaining, TIMEOUT_MINS
-from htcondor_es.convert_to_json import convert_to_json
-from htcondor_es.convert_to_json import convert_dates_to_millisecs
-from htcondor_es.convert_to_json import unique_doc_id
+from . import utils, convert
 
 
 class ListenAndBunch(multiprocessing.Process):
@@ -56,7 +52,7 @@ class ListenAndBunch(multiprocessing.Process):
         while True:
             try:
                 next_batch = self.input_queue.get(
-                    timeout=time_remaining(self.starttime - 5)
+                    timeout=utils.time_remaining(self.starttime - 5)
                 )
             except queue.Empty:
                 logging.warning("Closing listener before all schedds were processed")
@@ -93,68 +89,71 @@ class ListenAndBunch(multiprocessing.Process):
             if len(self.buffer) >= self.bunch_size:
                 self.output_queue.put(
                     self.buffer[: self.bunch_size],
-                    timeout=time_remaining(self.starttime),
+                    timeout=utils.time_remaining(self.starttime),
                 )
                 self.buffer = self.buffer[self.bunch_size :]
 
     def close(self):
         """Clear the buffer, send a poison pill and the total number of docs"""
         if self.buffer:
-            self.output_queue.put(self.buffer, timeout=time_remaining(self.starttime))
+            self.output_queue.put(
+                self.buffer, timeout=utils.time_remaining(self.starttime)
+            )
             self.buffer = []
 
         logging.warning("Closing listener, received %d documents total", self.count_in)
         # send back a poison pill
-        self.output_queue.put(None, timeout=time_remaining(self.starttime))
+        self.output_queue.put(None, timeout=utils.time_remaining(self.starttime))
         # send the number of total docs
-        self.output_queue.put(self.count_in, timeout=time_remaining(self.starttime))
+        self.output_queue.put(
+            self.count_in, timeout=utils.time_remaining(self.starttime)
+        )
 
 
 def query_schedd_queue(starttime, schedd_ad, queue, args):
     my_start = time.time()
     logging.info("Querying %s queue for jobs.", schedd_ad["Name"])
-    if time_remaining(starttime) < 10:
+    if utils.time_remaining(starttime) < 10:
         message = (
             "No time remaining to run queue crawler on %s; "
             "exiting." % schedd_ad["Name"]
         )
         logging.error(message)
-        send_email_alert(args.email_alerts, "spider_cms queue timeout warning", message)
+        utils.send_email_alert(
+            args.email_alerts, "spider_cms queue timeout warning", message
+        )
         return
 
     count_since_last_report = 0
     count = 0
     cpu_usage = resource.getrusage(resource.RUSAGE_SELF).ru_utime
-    queue.put(schedd_ad["Name"], timeout=time_remaining(starttime))
+    queue.put(schedd_ad["Name"], timeout=utils.time_remaining(starttime))
 
     schedd = htcondor.Schedd(schedd_ad)
     sent_warnings = False
     batch = []
     # Query for a snapshot of the jobs running/idle/held,
     # but only the completed that had changed in the last period of time.
-    _completed_since = starttime - (TIMEOUT_MINS + 1) * 60
-    query = ("""
+    _completed_since = starttime - (utils.TIMEOUT_MINS + 1) * 60
+    query = """
          (JobStatus < 3 || JobStatus > 4 
          || EnteredCurrentStatus >= %(completed_since)d)
-         """
-        % {"completed_since": _completed_since}
-    )
+         """ % {
+        "completed_since": _completed_since
+    }
     try:
         query_iter = schedd.xquery(requirements=query) if not args.dry_run else []
         for job_ad in query_iter:
             dict_ad = None
             try:
-                dict_ad = convert_to_json(
+                dict_ad = convert.to_json(
                     job_ad, return_dict=True, reduce_data=not args.keep_full_queue_data
                 )
             except Exception as e:
-                message = "Failure when converting document on %s queue: %s" % (
-                    schedd_ad["Name"],
-                    str(e),
-                )
+                message = f"Failure when converting document on {schedd_ad['Name']} queue: {e}"
                 logging.warning(message)
                 if not sent_warnings:
-                    send_email_alert(
+                    utils.send_email_alert(
                         args.email_alerts,
                         "spider_cms queue document conversion error",
                         message,
@@ -164,23 +163,23 @@ def query_schedd_queue(starttime, schedd_ad, queue, args):
             if not dict_ad:
                 continue
 
-            batch.append((unique_doc_id(dict_ad), dict_ad))
+            batch.append((convert.unique_doc_id(dict_ad), dict_ad))
             count += 1
             count_since_last_report += 1
 
             if not args.dry_run and len(batch) == args.query_queue_batch_size:
-                if time_remaining(starttime) < 10:
+                if utils.time_remaining(starttime) < 10:
                     message = (
                         "Queue crawler on %s has been running for "
                         "more than %d minutes; exiting"
-                        % (schedd_ad["Name"], TIMEOUT_MINS)
+                        % (schedd_ad["Name"], utils.TIMEOUT_MINS)
                     )
                     logging.error(message)
-                    send_email_alert(
+                    utils.send_email_alert(
                         args.email_alerts, "spider_cms queue timeout warning", message
                     )
                     break
-                queue.put(batch, timeout=time_remaining(starttime))
+                queue.put(batch, timeout=utils.time_remaining(starttime))
                 batch = []
                 if count_since_last_report >= 1000:
                     cpu_usage_now = resource.getrusage(resource.RUSAGE_SELF).ru_utime
@@ -208,21 +207,20 @@ def query_schedd_queue(starttime, schedd_ad, queue, args):
             "Failed to query schedd %s for jobs: %s", schedd_ad["Name"], str(e)
         )
     except Exception as e:
-        message = "Failure when processing schedd queue query on %s: %s" % (
-            schedd_ad["Name"],
-            str(e),
+        message = (
+            f"Failure when processing schedd queue query on {schedd_ad['Name']}: {e}"
         )
         logging.error(message)
-        send_email_alert(
+        utils.send_email_alert(
             args.email_alerts, "spider_cms schedd queue query error", message
         )
         traceback.print_exc()
 
     if batch:  # send remaining docs
-        queue.put(batch, timeout=time_remaining(starttime))
+        queue.put(batch, timeout=utils.time_remaining(starttime))
         batch = []
 
-    queue.put(schedd_ad["Name"], timeout=time_remaining(starttime))
+    queue.put(schedd_ad["Name"], timeout=utils.time_remaining(starttime))
     total_time = (time.time() - my_start) / 60.0
     logging.warning(
         "Schedd %-25s queue: response count: %5d; " "query time %.2f min; ",
@@ -239,7 +237,7 @@ def process_queues(schedd_ads, starttime, pool, args, metadata=None):
     Process all the jobs in all the schedds given.
     """
     my_start = time.time()
-    if time_remaining(starttime) < 10:
+    if utils.time_remaining(starttime) < 10:
         logging.warning("No time remaining to process queues")
         return
 
@@ -270,14 +268,16 @@ def process_queues(schedd_ads, starttime, pool, args, metadata=None):
         if args.dry_run or len(schedd_ads) == 0:
             break
 
-        if time_remaining(starttime) < 5:
+        if utils.time_remaining(starttime) < 5:
             logging.warning("Listener did not shut down properly; terminating.")
             listener.terminate()
             break
 
-        bunch = output_queue.get(timeout=time_remaining(starttime))
+        bunch = output_queue.get(timeout=utils.time_remaining(starttime))
         if bunch is None:  # swallow the poison pill
-            total_processed = int(output_queue.get(timeout=time_remaining(starttime)))
+            total_processed = int(
+                output_queue.get(timeout=utils.time_remaining(starttime))
+            )
             break
 
         if args.feed_es_for_queues and not args.read_only:
@@ -303,9 +303,9 @@ def process_queues(schedd_ads, starttime, pool, args, metadata=None):
     total_upload_time = 0
     total_queried = 0
     for name, future in futures:
-        if time_remaining(starttime, positive=False) > -20:
+        if utils.time_remaining(starttime, positive=False) > -20:
             try:
-                count = future.get(time_remaining(starttime) + 10)
+                count = future.get(utils.time_remaining(starttime) + 10)
                 if name == "UPLOADER_ES":
                     total_sent += count
                 else:
@@ -316,7 +316,7 @@ def process_queues(schedd_ads, starttime, pool, args, metadata=None):
             except multiprocessing.TimeoutError:
                 message = "Schedd %s queue timed out; ignoring progress." % name
                 logging.error(message)
-                send_email_alert(
+                utils.send_email_alert(
                     args.email_alerts, "spider_cms queue timeout warning", message
                 )
         else:
